pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "rensis"
version = "1.0.1"
dependencies = [
  "nsisunbz2",
  "pefile",
  "yara-python"
]
requires-python = ">=3.11"
